+++ packages/tkeel-console-plugin-tenant-users/src/pages/Users/components/ResetPasswordButton/index.tsx
@@ -1,6 +1,9 @@
 import { Text, useDisclosure } from '@chakra-ui/react';
 
-import { SetPasswordModal } from '@tkeel/console-business-components';
+import {
+  SetPasswordModal,
+  useSetPasswordUrl,
+} from '@tkeel/console-business-components';
 import { Alert, LinkButton } from '@tkeel/console-components';
 import {
   useGetResetPasswordKeyMutation,
@@ -36,6 +39,11 @@ export default function ResetPasswordButton({ data }: Props) {
     },
   });
 
+  const { isLoading: isSetPasswordUrlLoading, setPasswordUrl } =
+    useSetPasswordUrl({
+      data: { reset_key: resetData?.reset_key ?? '' },
+    });
+
   const handleConfirm = () => {
     mutate({});
   };
@@ -65,9 +73,10 @@ export default function ResetPasswordButton({ data }: Props) {
       {isSetPasswordModalOpen && (
         <SetPasswordModal
           isOpen={isSetPasswordModalOpen}
-          // url={url}
-          data={{ reset_key: resetData?.reset_key ?? '' }}
-          title="操作成功"
+          title="重置密码请求成功"
+          description="123"
+          url={setPasswordUrl}
+          isLoading={isSetPasswordUrlLoading}
           onClose={onSetPasswordModalClose}
         />
       )}
