+++ packages/tkeel-console-request-hooks/src/hooks/queries/useDeploymentConfigQuery/index.ts
@@ -32,13 +32,12 @@ export default function useDeploymentConfigQuery() {
   if (process.env.NODE_ENV === 'development') {
     try {
       const mockDocsBaseURL =
-        GLOBAL_PORTAL_CONFIG.mock?.documents?.baseURL || config.docsURL;
+        GLOBAL_PORTAL_CONFIG?.mock?.documents?.baseURL || config.docsURL;
       const mockConfig = merge({}, config, { docsURL: mockDocsBaseURL });
 
       return { config: mockConfig, ...result };
-    } catch (error) {
-      // eslint-disable-next-line no-console
-      console.error(error);
+    } catch {
+      //
     }
   }
